+++ src/main/java/br/apae/ged/exceptions/AlunoNaoEncontradoException.java
@@ -1,4 +1,4 @@
-package br.apae.ged.exception;
+package br.apae.ged.exceptions;
 
 public class AlunoNaoEncontradoException extends RuntimeException {
     public AlunoNaoEncontradoException(Long id) {
