+++ src/main/java/br/apae/ged/services/AlunoService.java
@@ -1,9 +1,8 @@
-package br.apae.ged.service;
+package br.apae.ged.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class AlunoService {
